Created FetchRequest
